+++ src/main/java/woooo_app/woooo/Navigation.kt
@@ -7,10 +7,11 @@ import androidx.compose.runtime.Composable
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.annotation.RootNavGraph
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
+import woooo_app.woooo.feature.auth.screen.SignUpView
 import com.wgroup.woooo_app.woooo.feature.home.screen.DashboardView
 import com.wgroup.woooo_app.woooo.feature.home.screen.HomePage
 import com.wgroup.woooo_app.woooo.feature.mining.views.MiningMainView
-import com.wgroup.woooo_app.woooo.feature.profile.views.UpdateProfileView
+import woooo_app.woooo.feature.profile.views.UpdateProfileView
 import com.wgroup.woooo_app.woooo.feature.settings.views.ApplicationMainView
 import com.wgroup.woooo_app.woooo.feature.settings.views.SettingMainView
 import com.wgroup.woooo_app.woooo.feature.settings.views.account.AccountMainView
@@ -22,21 +23,20 @@ import com.wgroup.woooo_app.woooo.feature.settings.views.application.LanguageVie
 import com.wgroup.woooo_app.woooo.feature.settings.views.application.SoundAndVibrationView
 import com.wgroup.woooo_app.woooo.feature.wallet.views.SendCurrencyView
 import com.wgroup.woooo_app.woooo.feature.wallet.views.TransactionsView
-import woooo_app.woooo.feature.auth.screen.SignUpView
 import woooo_app.woooo.feature.wallet.views.WalletMainView
-
+@RootNavGraph(start = true)
 @Destination
 @Composable
 fun LoginScreen(navigator: DestinationsNavigator) {
     LoginView(navigator)
 }
 
+
 @Destination
 @Composable
 fun HomeScreen(navigator: DestinationsNavigator) {
     HomePage(navigator = navigator)
 }
-@RootNavGraph(start = true)
 
 @Destination
 @Composable