+++ src/main/java/woooo_app/woooo/data/datasource/remote/auth/AuthApiService.kt
@@ -4,6 +4,7 @@ import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
+import retrofit2.http.Query
 import retrofit2.http.QueryMap
 import woooo_app.woooo.data.models.auth.ConfirmAccountModel
 import woooo_app.woooo.data.models.auth.LoginModel
@@ -18,8 +19,11 @@ interface AuthApiService {
 //        const val BASE_URL = "http://192.168.1.18"
     }
 
-    @POST("/api/Auth/LoginWithEmail")
-    suspend fun login(@Body user: LoginRequestParams): Response<LoginModel>
+    @POST("/api/Auth/login")
+    suspend fun login(
+        @Query("isLoginWithEmail") isLoginWithEmail: Boolean,
+        @Body user: LoginRequestParams
+    ): Response<LoginModel>
 
     @POST("/api/Auth/SignUp")
     suspend fun signUp(@Body user: SignUpRequestModel): Response<SignUpModel>
@@ -32,6 +36,5 @@ interface AuthApiService {
 
     @POST("/api/Auth/forgot-password")
     suspend fun forgotPassword(@QueryMap params: String): Response<ResendCodeModel>
-
 
 }