+++ src/main/java/woooo_app/woooo/data/models/auth/requestmodels/LoginRequestParams.kt
@@ -2,6 +2,9 @@ package woooo_app.woooo.data.models.auth.requestmodels
 
 data class LoginRequestParams(
     val email: String="",
-    val deviceName: String="",
-    val password: String=""
+    val PhoneNumber: String="",
+    val password: String="",
+    val deviceId: String="",
+    val ipAddress: String="",
+    val userAgent: String="",
 )